psa-arch-tests: Allow overriding of toolchain file
The INHERIT.cmake toolchain file that upstream intends us to use is
not flexible enough. It does not permit overriding of CMAKE_C_FLAGS
for instance as GNUARM.cmake does.

So we allow the file to be overridden such that we can have full
flexibility in customizing the toolchain.

This patch is temporary until upstream finds a better solution for
toolchain configuration.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>